requirements.txt
Django>=4.2.1,<4.3
psycopg2>=2.8.6,<2.9
djangorestframework
celery[redis]
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==4.5.5
uWSGI>=2.0.19.1,<2.1
django-cors-headers
